+++ Circulo de seguridad/Entidades/Localizacion.cs
@@ -8,8 +8,8 @@ namespace Circulo_de_seguridad.Entidades
     public class Localizacion
     {
         public int Id { get; set; }
-        public int CoordenadaX { get; set; }
-        public int CoordenadaY { get; set; }
+        public float CoordenadaX { get; set; }
+        public float CoordenadaY { get; set; }
         public int UsuarioId { get; set; }
         public DateTime Fecha { get; set; }
         public Usuario Usuario { get; set; }
